requirements.txt
streamlit
google-generativeai
python-dotenv
youtube_transcript_api==0.6.3